my_agent/requirements.txt
langgraph==0.2.69
langchain-deepseek-official==0.1.0
langgraph-cli[inmem]